requirements.txt
Flask
Flask-Session
Flask-CORS
bcrypt
pandas
openpyxl
paramiko
numexpr
numpy
python-dotenv
pyodbc
werkzeug